requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
mlflow
jupyter
dvc
optuna
hydra-core
hydra-optuna-sweeper